requirements.txt
# Core dependencies
streamlit==1.32.0
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.1
PyYAML==6.0.1
opencv-python==4.8.1.78
scikit-learn==1.3.0
deepface==0.0.79

# Face detection
retinaface==0.0.13

# Head pose
opencv-contrib-python==4.8.1.78

# Database
sqlite3

# API integration
groq==0.4.1

# Visualization
matplotlib==3.7.3
plotly==5.18.0